+++ elmyra.ip.access.epo/CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - api: add a little cql smartness: wrap cql query string with
        quotes if query contains spaces and is still unquoted
+- api: enhance image information, publish via endpoint
 
 0.3.0
 =====